rename server/index.js to server/respond.js
